requirements.txt
# Database
pandas
psycopg2-binary
sqlalchemy

# Profiling report
reportlab